Kind of uniform point sampling (doesn't work as expected)

+++ point_sampler.hpp
@@ -16,5 +16,5 @@ namespace clouds
     extern Eigen::Vector3f sampleTrianglePoint(const Eigen::Vector3f& a, const Eigen::Vector3f& b, const Eigen::Vector3f& c, float r1, float r2);
 
     // TODO : Use samples number per unit of area
-    extern pcl::PointCloud<pcl::PointXYZLNormal>::Ptr sampleMesh(const stl_reader::StlMesh<float, unsigned>& mesh, size_t samples);
+    extern pcl::PointCloud<pcl::PointXYZLNormal>::Ptr sampleMesh(const stl_reader::StlMesh<float, unsigned>& mesh, float samples);
 }
